+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Common
   
   * Fixed bug in StepSequence::getMaxSteps(): [#305](https://github.com/personalrobotics/aikido/pull/305)
+  * Fixed bug in StepSequence iterator: [#320](https://github.com/personalrobotics/aikido/pull/320)
 
 * State Space
 
@@ -18,6 +19,7 @@
 * Planner
 
   * Added parabolic timing for linear spline [#302](https://github.com/personalrobotics/aikido/pull/302)
+  * Fixed step sequence iteration in VPF: [#303](https://github.com/personalrobotics/aikido/pull/303)
 
 ### 0.2.0 (2018-01-09)
 
@@ -38,7 +40,7 @@
 * Planner
 
   * Added World class: [#243](https://github.com/personalrobotics/aikido/pull/243), [#252](https://github.com/personalrobotics/aikido/pull/252), [#265](https://github.com/personalrobotics/aikido/pull/265)
-  * Added vector field planner [#246](https://github.com/personalrobotics/aikido/pull/246), [#262](https://github.com/personalrobotics/aikido/pull/262), [#268](https://github.com/personalrobotics/aikido/pull/268)
+  * Added vector field planner: [#246](https://github.com/personalrobotics/aikido/pull/246), [#262](https://github.com/personalrobotics/aikido/pull/262), [#268](https://github.com/personalrobotics/aikido/pull/268)
 
 * RViz